Proto filtering of metric values.
Stripped out the old pmclient_fg.c code and added code to have a
single predicate and multiple metrics that are filterd by the
predicate.  The printout is primitive, printing out all the predicate
values and each of the instances in the metrics that have the
predicate evaluate to some value greate than 0.  The predicate is
specified with the "-f" or "--filter" option and each the predicates
is specified with the "-m" or "--metric" option.  If no predicate is
specified, then all instances of the metrics are printed.

As a quick test could use the following to get utime and stime for
non-root processes:

./pmpredicate -f proc.id.uid -m proc.psinfo.utime -m proc.psinfo.stime

The following would print out utime and stime metrics for all processes:

./pmpredicate -f proc.id.uid -m proc.psinfo.utime -m proc.psinfo.stime

There are a number of things that need to be addressed in the code:

-Improve checks that the predicate is a numerical value
-Verify that the instances of the predicate match up with the metrics
-More flexibility in the metrics, right now only numerical ones are handled
-Handle cases where the number of instances can change over time