Improve the predicate search and matching
The matching previously made an incorrect assumption that the same
positions in the arrays were for the same instances.  The predicate
and metric could have data for difference instances for the same
index.  The matching was revised to use the instance number rather
than raw index in the array.  This should be more robust for metrics
where the set of instances could quickly change resulting in the
predicate and metrics have data for the same instances at different
indices in the respective arrays.