Add an option to limit results to top n with highest predicate values
The predicate in some cases may be a scalar value rather than just a
binary true/false value.  Larger scalar values could indicate greater
resource usage such as amount of processor time or resident memory
used.  When the predicate is a scalar value it might be useful to
limit the results to a subset of the largest predicate values, for
example provide a list of ten most cpu intensive tasks rather than
listing every process that used processor time.  This new option makes
it easy to implement this filtering.