Add code to eliminate '.' in the metric names.
For the PCP Metric Name Space (PMNS) '.' has special meaning
indicating things lower in the tree.  Right now pmpredicate just
flattens the metrics in one level.  Having the '.' in the metric name
confuses the PCP PMNS code.  For the time being mangle the names to
remove any '.'.  In the longer term would probably prefer to preserve
the PMNS tree structure of the original metric.