Use a temporary file and rename to ensure that the data.json atomically updated
There was a possibility that the json pmda would be reading the
data.json while pmdapredicate was overwriting the file.  pmpredicate
now writes data.json to a temporary file and does a file rename to
atomically update the data.json file.  This will allow the json pmda
to finish reading the old file data and avoid any issues with
corruption because it had mixture of old and new data.