Avoid pmapi fetch converting counters into rates
pmpredicate should be preserving the filtered metric as much as
possible.  The pmapi fetch was converting counters into rates and then
what every reading those counters was generating a rate from the rate
value.  The code passes in "instant" to the pmapi fetch avoid the
double rate conversion of the metric.