Add a testcase for pmpredicate to pcp qa testsuite
This is a simple test to make sure that two separate pmpredicate
instances can run concurrently and provide some meaningful data.  This
is still a work in progress.  The PCP caching of instance names seems
to be causing some problems and pmval will return following error
message if the instance changes in the json data:

pmval: pmFetch: Unknown or illegal instance identifier