Place the temporary data.json file in the same directory as data.json
The -d (directory) option was not placing the data.json.tmp in the
same directory as the data.json and metadata.json files.  This was
causing the rename operation to fail and pmda json would never present
the information metrics because the data.json files were not found.
The data.json.tmp is now in the same directory as the other.