Use instance names rather that instance numbers for array keys
To make the data.json a bit easier to read changed the array key from
the instance numbers to the instance name.  This allows someone
examining the data.json file to know what instance the metric is
associated with without having to looking up the mapping of number to
instance name.  However, this will encounter a bit more overhead because
all the strings that need to be fetched for each of the instances.