Check to make sure that the predicate and metrics have the same instance domain
The instances in the predicate should line with the instances in the
metrics. All the fetched metrics should have the same instance domain
for the instances to match up.  If they are not in the same instance
domain, the instances are not going to match up.  pmpredicate now
flags this instance domain mismatch as an error.