Improve handling of predicate and rank options for configuration files
Want to make sure that any addition steps required for an option are
done regardless whether the option is specified on the command line or
in a configuration file.  Moved those adjustments and checks out of
the options function.  Some of the steps may need to take place before
the establishment of a PMAPI context so an adjust_opts function has
been created for these cases.