Simplify the code extracting metric instances with true predicates
The code to get a set of instances with true predicates and print them
out was a bit convoluted.  Reordered the operations to move the rank
subset operation earlier to subset the data as quickly as possible and
reduce the amount of data moved around.  Removed some of the nested
loops. Thus, the steps are now:

-Get a list of instances from the predicate metric
-Reduce the list to a maximum number of instances if rank option used
-Make a dictionary of the predicate instances
-For each metric find the instance that match the predicate instances
-For each predicate instance print out the matching metric instances

Also added a validate_config check to make sure that the predicate
metric has instances and is not a scalar.